ANY1 v3 update
- emulator

+++ v3/software/FEM/source/Disassem.h
@@ -1,4 +1,4 @@
 #pragma once
 
-std::string Disassem(std::string, std::string, unsigned int, unsigned int *);
-std::string Disassem(unsigned int, unsigned int, unsigned int *);
+std::string Disassem(std::string, std::string, uint64_t, unsigned int *);
+std::string Disassem(unsigned int, uint64_t, unsigned int *);
